+++ R/helpers_easystats.R
@@ -4,8 +4,9 @@
 #' @inheritParams parameters::model_parameters
 #'
 #' @importFrom parameters model_parameters standardize_names
-#' @importFrom dplyr select matches rename_all recode contains
+#' @importFrom dplyr select matches rename_all recode contains filter bind_cols
 #' @importFrom tidyr fill
+#' @importFrom performance r2_bayes
 #'
 #' @examples
 #' model <- lm(mpg ~ wt + cyl, data = mtcars)
@@ -13,12 +14,32 @@
 #' @export
 
 tidy_model_parameters <- function(model, ...) {
-  parameters::model_parameters(model, verbose = FALSE, ...) %>%
+  stats_df <- parameters::model_parameters(model, verbose = FALSE, ...) %>%
     dplyr::select(-dplyr::matches("Difference")) %>%
     parameters::standardize_names(style = "broom") %>%
     dplyr::rename_all(.funs = dplyr::recode, "bayes.factor" = "bf10") %>%
     tidyr::fill(dplyr::matches("^prior|^bf"), .direction = "updown") %>%
     as_tibble(.)
+
+  # ------------------------ Bayesian ANOVA designs -------------------------
+
+  if ("method" %in% names(stats_df)) {
+    if (stats_df$method[[1]] == "Bayes factors for linear models") {
+      # dataframe with posterior estimates for R-squared
+      df_r2 <- performance::r2_bayes(model, average = TRUE, ci = stats_df$conf.level[[1]]) %>%
+        as_tibble(.) %>%
+        parameters::standardize_names(style = "broom") %>%
+        dplyr::rename_with(.fn = ~ paste0("r2.", .x), .cols = dplyr::matches("^conf|^comp"))
+
+      # for within-subjects design, retain only marginal component
+      if ("r2.component" %in% names(df_r2)) df_r2 %<>% dplyr::filter(r2.component == "conditional")
+
+      # combine everything
+      stats_df %<>% dplyr::bind_cols(df_r2)
+    }
+  }
+
+  stats_df
 }